+++ e2e/nx-aws-cdk-e2e/tests/aws-cdk.spec.ts
@@ -1,10 +1,10 @@
-import { checkFilesExist, ensureNxProject, readJson, runNxCommandAsync, uniq } from '@nrwl/nx-plugin/testing';
+import { checkFilesExist, ensureNxProject, readJson, runNxCommandAsync, uniq } from '@nx/plugin/testing';
+
+beforeAll(async () => {
+  ensureNxProject('@codebrew/nx-aws-cdk', 'dist/packages/nx-aws-cdk');
+});
 
 describe('nx-aws-cdk e2e', () => {
-  beforeAll(() => {
-    ensureNxProject('@codebrew/nx-aws-cdk', 'dist/packages/nx-aws-cdk');
-  });
-
   it('should create aws-cdk', async () => {
     const plugin = uniq('nx-aws-cdk');
 
@@ -16,7 +16,7 @@ describe('nx-aws-cdk e2e', () => {
       const plugin = uniq('nx-aws-cdk');
 
       await runNxCommandAsync(`generate @codebrew/nx-aws-cdk:application ${plugin} --directory subdir`);
-      expect(() => checkFilesExist(`apps/subdir/${plugin}/src/main.ts`)).not.toThrow();
+      expect(() => checkFilesExist(`subdir/${plugin}/src/main.ts`)).not.toThrow();
     }, 120000);
   });
 
@@ -25,8 +25,8 @@ describe('nx-aws-cdk e2e', () => {
       const plugin = uniq('nx-aws-cdk');
 
       await runNxCommandAsync(`generate @codebrew/nx-aws-cdk:application ${plugin} --tags e2etag,e2ePackage`);
-      const nxJson = readJson('nx.json');
-      expect(nxJson.projects[plugin].tags).toEqual(['e2etag', 'e2ePackage']);
+      const projectJson = readJson(`${plugin}/project.json`);
+      expect(projectJson.tags).toEqual(['e2etag', 'e2ePackage']);
     }, 120000);
   });
 });