+++ jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  projects: ['<rootDir>/packages/nx-aws-cdk', '<rootDir>/e2e/nx-aws-cdk-e2e'],
-};