+++ packages/nx-aws-cdk/src/executors/destroy/destroy.ts
@@ -1,9 +1,10 @@
 import * as path from 'path';
-import { ExecutorContext } from '@nrwl/tao/src/shared/workspace';
 
 import { DestroyExecutorSchema } from './schema';
-import { createCommand, parseArgs, runCommandProcess } from '../../utils/executor.util';
+import { createCommand, parseArgs } from '../../utils/executor.util';
 import { ParsedExecutorInterface } from '../../interfaces/parsed-executor.interface';
+import type { ExecutorContext } from '@nx/devkit';
+import { runCommandProcess } from '@codebrew/utils';
 
 export interface ParsedDestroyExecutorOption extends ParsedExecutorInterface {
   parseArgs?: Record<string, string>;
@@ -35,7 +36,7 @@ function normalizeOptions(options: DestroyExecutorSchema, context: ExecutorConte
     stacks = options.stacks;
   }
 
-  const { sourceRoot, root } = context?.workspace?.projects[context.projectName];
+  const { sourceRoot, root } = context.workspace.projects[context.projectName];
 
   return {
     ...options,