+++ packages/nx-aws-cdk/src/generators/init/init.ts
@@ -1,8 +1,8 @@
-import { addDependenciesToPackageJson, convertNxGenerator, formatFiles, GeneratorCallback, Tree } from '@nrwl/devkit';
-import { jestInitGenerator } from '@nrwl/jest';
+import { addDependenciesToPackageJson, convertNxGenerator, formatFiles, GeneratorCallback, Tree } from '@nx/devkit';
+import { jestInitGenerator } from '@nx/jest';
 
 import { InitGeneratorSchema } from './schema';
-import { CDK_ESLINT_VERSION, CDK_VERSION } from '../../utils/cdk-shared';
+import { CDK_CONSTRUCTS_VERSION, CDK_ESLINT_VERSION, CDK_VERSION } from '../../utils/cdk-shared';
 
 function normalizeOptions(schema: InitGeneratorSchema) {
   return {
@@ -22,8 +22,9 @@ export async function initGenerator(host: Tree, options: InitGeneratorSchema) {
   const installTask = addDependenciesToPackageJson(
     host,
     {
-      '@aws-cdk/core': CDK_VERSION,
-      '@aws-cdk/assert': CDK_VERSION,
+      'aws-cdk': CDK_VERSION,
+      'aws-cdk-lib': CDK_VERSION,
+      constructs: CDK_CONSTRUCTS_VERSION,
     },
     {
       'eslint-plugin-cdk': CDK_ESLINT_VERSION,