+++ packages/nx-aws-cdk/src/utils/cdk-shared.ts
@@ -1,2 +1,3 @@
-export const CDK_VERSION = '^1.127.0';
-export const CDK_ESLINT_VERSION = '^1.3.0';
+export const CDK_VERSION = '^2.77.0';
+export const CDK_ESLINT_VERSION = '^1.8.0';
+export const CDK_CONSTRUCTS_VERSION = '^10.2.10';