+++ packages/nx-ncc/src/interfaces/base-executor-schema.interface.ts
@@ -1,0 +1,11 @@
+import type { AssetGlob } from '@nx/js/src/utils/assets/assets';
+
+export interface BaseExecutorSchemaInterface {
+  main: string;
+  sourceMap?: boolean;
+  minify?: boolean;
+  quiet?: boolean;
+  target?: string;
+  externalDependencies?: string[];
+  assets?: (string | AssetGlob)[];
+}