+++ packages/nx-ncc/src/interfaces/normalized-executor-schema.interface.ts
@@ -1,0 +1,8 @@
+import { BaseExecutorSchemaInterface } from './base-executor-schema.interface';
+
+export type NormalizedExecutorSchemaInterface<T extends BaseExecutorSchemaInterface = BaseExecutorSchemaInterface> =
+  T & {
+    root: string;
+    sourceRoot: string;
+    projectRoot: string;
+  };