+++ packages/nx-aws-cdk/CHANGELOG.md
@@ -2,6 +2,52 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## 2.0.0 (2024-08-27)
+
+## 1.0.0 (2024-08-27)
+
+- chore: migrate 14.8.6 ([5ea5d38](https://github.com/codebrewlab/nx-plugins/commit/5ea5d38))
+- chore: nx migrate 13.10.6 ([08bf0e0](https://github.com/codebrewlab/nx-plugins/commit/08bf0e0))
+- chore: nx migrate 15.9.7 ([820fc52](https://github.com/codebrewlab/nx-plugins/commit/820fc52))
+- chore: nx migrate 16.10.0 ([5477ff8](https://github.com/codebrewlab/nx-plugins/commit/5477ff8))
+- chore: nx migrate 17.3.2 ([7925d6d](https://github.com/codebrewlab/nx-plugins/commit/7925d6d))
+- chore(aws-cdk): add license info ([6385c22](https://github.com/codebrewlab/nx-plugins/commit/6385c22))
+- chore(release): 0.1.0 ([6a560cd](https://github.com/codebrewlab/nx-plugins/commit/6a560cd))
+- chore(release): 0.1.1 ([dcecfb7](https://github.com/codebrewlab/nx-plugins/commit/dcecfb7))
+- chore(release): 1.0.0 ([98c9de1](https://github.com/codebrewlab/nx-plugins/commit/98c9de1))
+- chore(release): 1.0.1 ([a76cb78](https://github.com/codebrewlab/nx-plugins/commit/a76cb78))
+- chore(release): 1.0.2 ([7f8998d](https://github.com/codebrewlab/nx-plugins/commit/7f8998d))
+- chore(release): 1.0.3 ([e68a947](https://github.com/codebrewlab/nx-plugins/commit/e68a947))
+- fix(aws-cdk): executor stdin fix ([b5ef5c7](https://github.com/codebrewlab/nx-plugins/commit/b5ef5c7))
+- fix(aws-cdk): fix nx linter (#2) ([58673ae](https://github.com/codebrewlab/nx-plugins/commit/58673ae)), closes [#2](https://github.com/codebrewlab/nx-plugins/issues/2)
+- fix(aws-cdk): miss await in destroy executor ([23e1815](https://github.com/codebrewlab/nx-plugins/commit/23e1815))
+- fix(nx-aws-cdk): fix destroy executor (#4) ([2f48bc0](https://github.com/codebrewlab/nx-plugins/commit/2f48bc0)), closes [#4](https://github.com/codebrewlab/nx-plugins/issues/4)
+- test(aws-cdk): logger info to debug ([be3b4c7](https://github.com/codebrewlab/nx-plugins/commit/be3b4c7))
+- feat(aws-cdk): deploy, destroy executor add option ([b48e374](https://github.com/codebrewlab/nx-plugins/commit/b48e374))
+- feat(aws-cdk): generator, executor complete ([84dd16a](https://github.com/codebrewlab/nx-plugins/commit/84dd16a))
+
+## 1.0.0 (2024-08-27)
+
+- chore: migrate 14.8.6 ([5ea5d38](https://github.com/codebrewlab/nx-plugins/commit/5ea5d38))
+- chore: nx migrate 13.10.6 ([08bf0e0](https://github.com/codebrewlab/nx-plugins/commit/08bf0e0))
+- chore: nx migrate 15.9.7 ([820fc52](https://github.com/codebrewlab/nx-plugins/commit/820fc52))
+- chore: nx migrate 16.10.0 ([5477ff8](https://github.com/codebrewlab/nx-plugins/commit/5477ff8))
+- chore: nx migrate 17.3.2 ([7925d6d](https://github.com/codebrewlab/nx-plugins/commit/7925d6d))
+- chore(aws-cdk): add license info ([6385c22](https://github.com/codebrewlab/nx-plugins/commit/6385c22))
+- chore(release): 0.1.0 ([6a560cd](https://github.com/codebrewlab/nx-plugins/commit/6a560cd))
+- chore(release): 0.1.1 ([dcecfb7](https://github.com/codebrewlab/nx-plugins/commit/dcecfb7))
+- chore(release): 1.0.0 ([98c9de1](https://github.com/codebrewlab/nx-plugins/commit/98c9de1))
+- chore(release): 1.0.1 ([a76cb78](https://github.com/codebrewlab/nx-plugins/commit/a76cb78))
+- chore(release): 1.0.2 ([7f8998d](https://github.com/codebrewlab/nx-plugins/commit/7f8998d))
+- chore(release): 1.0.3 ([e68a947](https://github.com/codebrewlab/nx-plugins/commit/e68a947))
+- fix(aws-cdk): executor stdin fix ([b5ef5c7](https://github.com/codebrewlab/nx-plugins/commit/b5ef5c7))
+- fix(aws-cdk): fix nx linter (#2) ([58673ae](https://github.com/codebrewlab/nx-plugins/commit/58673ae)), closes [#2](https://github.com/codebrewlab/nx-plugins/issues/2)
+- fix(aws-cdk): miss await in destroy executor ([23e1815](https://github.com/codebrewlab/nx-plugins/commit/23e1815))
+- fix(nx-aws-cdk): fix destroy executor (#4) ([2f48bc0](https://github.com/codebrewlab/nx-plugins/commit/2f48bc0)), closes [#4](https://github.com/codebrewlab/nx-plugins/issues/4)
+- test(aws-cdk): logger info to debug ([be3b4c7](https://github.com/codebrewlab/nx-plugins/commit/be3b4c7))
+- feat(aws-cdk): deploy, destroy executor add option ([b48e374](https://github.com/codebrewlab/nx-plugins/commit/b48e374))
+- feat(aws-cdk): generator, executor complete ([84dd16a](https://github.com/codebrewlab/nx-plugins/commit/84dd16a))
+
 ## [1.0.3](https://github.com/codebrewlab/nx-plugins/compare/v1.0.2...v1.0.3) (2022-01-28)
 
 ### Bug Fixes